Fix highlight: force background override and fix remove cls
The highlight feature works adding and removing a CSS class that sets a
background color.

This commits add "!important" to override any other background color.

The other fix is when we remove the CSS class from a previous
highlighted selection. Incorrectly we were iterating over the current
selection instead of the previous one.