Fix RC and corruption in `highlight()`
The `highlight()` function was removing the previous highlighted elements.

First, two consecutive calls to `highlight()` will make the second to
clear the first one (good) but the first one may still be re-activated due
the `setTimeout(toggle_highlight())`.

Fixed with the cancellation of any previous timeout callback.

Second, the `getElementsByClassName()` returns a live array: if we remove
the class from an element, the array is modified so a for-loop will not
work.

Fixed with a while-loop.